chore: move icon resolver code to dock
