+++ crates/wayle-shell/src/shell/bar/dropdowns/registry.rs
@@ -201,6 +201,7 @@ impl DropdownInstance {
     }
 
     fn apply_style(&self, style: &DropdownStyle) {
+        self.popover.set_opacity(style.opacity);
         self.popover.set_autohide(style.autohide);
         if style.shadow_enabled {
             self.popover.add_css_class("shadow");
